switched from float to double

+++ raytracer/Intersection.h
@@ -5,8 +5,8 @@ class Object;
 class Ray;
 
 struct Intersection {
-    Intersection(float dist, Ray* ray, Object* object): dist(dist), ray(ray), object(object) {};
-    float dist;
+    Intersection(double dist, Ray* ray, Object* object): dist(dist), ray(ray), object(object) {};
+    double dist;
     Object* object;
     Ray* ray;
 };
