+++ raytracer/objects/Object.h
@@ -12,7 +12,7 @@ public:
         , inverseTransform(1.f)
     {}
 
-    virtual float intersect(const Ray* ray) const { return 0; }
+    virtual double intersect(const Ray* ray) const { return 0; }
 
     virtual glm::vec3 getNormal(glm::vec4 point) = 0;
