+++ raytracer/objects/Triangle.h
@@ -11,7 +11,7 @@ class Triangle : public Object {
 public:
     Triangle(const vec3* v1, const vec3* v2, const vec3* v3);
 
-    float intersect(const Ray* ray) const override;
+    double intersect(const Ray* ray) const override;
 
     glm::vec3 getNormal(glm::vec4 point) override;
